Update to NLog 4.5-alpha02
