+++ src/NLog.Targets.Seq/LogEventInfoFormatter.cs
@@ -13,43 +13,17 @@
 // limitations under the License.
 
 using NLog.StructuredEvents;
+using NLog.StructuredEvents.Parts;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading;
 
 namespace NLog.Targets.Seq
 {
     static class LogEventInfoFormatter
     {
-        static readonly IDictionary<Type, Action<object, TextWriter>> LiteralWriters;
         const string InfoLevel = "Info";
-
-        static LogEventInfoFormatter()
-        {
-            LiteralWriters = new Dictionary<Type, Action<object, TextWriter>>
-            {
-                { typeof(bool), (v, w) => WriteBoolean((bool)v, w) },
-                { typeof(char), (v, w) => WriteString(((char)v).ToString(CultureInfo.InvariantCulture), w) },
-                { typeof(byte), WriteToString },
-                { typeof(sbyte), WriteToString },
-                { typeof(short), WriteToString },
-                { typeof(ushort), WriteToString },
-                { typeof(int), WriteToString },
-                { typeof(uint), WriteToString },
-                { typeof(long), WriteToString },
-                { typeof(ulong), WriteToString },
-                { typeof(float), WriteToString },
-                { typeof(double), WriteToString },
-                { typeof(decimal), WriteToString },
-                { typeof(string), (v, w) => WriteString((string)v, w) },
-                { typeof(DateTime), (v, w) => WriteDateTime((DateTime)v, w) },
-                { typeof(DateTimeOffset), (v, w) => WriteOffset((DateTimeOffset)v, w) },
-            };
-        }
 
         public static void ToCompactJson(IEnumerable<LogEventInfo> logEvents, TextWriter output, IList<SeqPropertyItem> properties)
         {
@@ -66,10 +40,6 @@ namespace NLog.Targets.Seq
             if (output == null) throw new ArgumentNullException(nameof(output));
             if (properties == null) throw new ArgumentNullException(nameof(properties));
 
-            // Without this call, logEventInfo.Properties behaves erratically when templates are used; stll trying
-            // to track down the cause (needs some debugging into NLog).
-            logEvent.FormattedMessage?.ToString();
-
             output.Write("{\"@t\":\"");
             output.Write(logEvent.TimeStamp.ToUniversalTime().ToString("O"));
 
@@ -80,7 +50,7 @@ namespace NLog.Targets.Seq
                 if (logEvent.Message != null)
                 {
                     message = logEvent.Message;
-                    template = TemplateParser.Parse(logEvent.Message);
+                    template = logEvent.GetmessageTemplate();
                 }
             }
             catch (TemplateParserException) { }
@@ -88,37 +58,47 @@ namespace NLog.Targets.Seq
             if (message != null || (message == null && logEvent.FormattedMessage == null))
             {
                 output.Write("\",\"@mt\":");
-                WriteString(message ?? "(No message)", output);
+                JsonWriter.WriteString(message ?? "(No message)", output);
             }
             else
             {
                 output.Write("\",\"@m\":");
-                WriteString(logEvent.FormattedMessage, output);
+                JsonWriter.WriteString(logEvent.FormattedMessage, output);
             }
 
-            var tokensWithFormat = template.Holes.Where(h => h.Format != null);
-
-            // Better not to allocate an array in the 99.9% of cases where this is false
-            // ReSharper disable once PossibleMultipleEnumeration
-            if (tokensWithFormat.Any())
+            if (template != null)
             {
-                output.Write(",\"@r\":[");
-                var delim = "";
-                foreach (var r in tokensWithFormat)
+                List<Hole> tokensWithFormat = null;
+                for (var i = 0; i < template.Holes.Length; ++i)
                 {
-                    output.Write(delim);
-                    delim = ",";
-                    var space = new StringWriter();
-                    var formatString = "{0:" + r.Format + "}";
-
-                    if (r.Name != null && logEvent.Properties != null && logEvent.Properties.ContainsKey(r.Name))
-                        space.Write(formatString, logEvent.Properties[r.Name]);
-                    else if (logEvent.Parameters != null && logEvent.Parameters.Length >= r.Index)
-                        space.Write(formatString, logEvent.Parameters[r.Index]);
-
-                    WriteString(space.ToString(), output);
+                    var hole = template.Holes[i];
+                    if (hole.Format != null)
+                    {
+                        tokensWithFormat = tokensWithFormat ?? new List<Hole>();
+                        tokensWithFormat.Add(hole);
+                    }
                 }
-                output.Write(']');
+
+                if (tokensWithFormat != null)
+                {
+                    output.Write(",\"@r\":[");
+                    var delim = "";
+                    foreach (var r in tokensWithFormat)
+                    {
+                        output.Write(delim);
+                        delim = ",";
+                        var space = new StringWriter();
+                        var formatString = "{0:" + r.Format + "}";
+
+                        if (r.Name != null && logEvent.Properties != null && logEvent.Properties.ContainsKey(r.Name))
+                            space.Write(formatString, logEvent.Properties[r.Name]);
+                        else if (logEvent.Parameters != null && logEvent.Parameters.Length >= r.Index)
+                            space.Write(formatString, logEvent.Parameters[r.Index]);
+
+                        JsonWriter.WriteString(space.ToString(), output);
+                    }
+                    output.Write(']');
+                }
             }
 
             if (logEvent.Level.Name != InfoLevel)
@@ -131,7 +111,7 @@ namespace NLog.Targets.Seq
             if (logEvent.Exception != null)
             {
                 output.Write(",\"@x\":");
-                WriteString(logEvent.Exception.ToString(), output);
+                JsonWriter.WriteString(logEvent.Exception.ToString(), output);
             }
 
             var seenKeys = new HashSet<string>();
@@ -145,7 +125,7 @@ namespace NLog.Targets.Seq
                 seenKeys.Add(name);
 
                 output.Write(',');
-                WriteString(name, output);
+                JsonWriter.WriteString(name, output);
                 output.Write(':');
 
                 var stringValue = property.Value.Render(logEvent);
@@ -154,16 +134,16 @@ namespace NLog.Targets.Seq
                     decimal numberValue;
                     if (decimal.TryParse(stringValue, out numberValue))
                     {
-                        WriteLiteral(numberValue, output);
+                        JsonWriter.WriteLiteral(numberValue, output);
                         continue;
                     }
                 }
 
-                WriteString(stringValue, output);
+                JsonWriter.WriteString(stringValue, output);
             }
 
-            if (logEvent.Message != null &&
-                logEvent.Message.Contains("{0") &&
+            if (template != null &&
+                template.IsPositional &&
                 logEvent.Parameters != null)
             {
                 for (var i = 0; i < logEvent.Parameters.Length; ++i)
@@ -175,9 +155,9 @@ namespace NLog.Targets.Seq
                     seenKeys.Add(name);
 
                     output.Write(',');
-                    WriteString(name, output);
+                    JsonWriter.WriteString(name, output);
                     output.Write(':');
-                    WriteLiteral(logEvent.Parameters[i], output);
+                    JsonWriter.WriteLiteral(logEvent.Parameters[i], output);
                 }
             }
 
@@ -192,9 +172,9 @@ namespace NLog.Targets.Seq
                     seenKeys.Add(name);
 
                     output.Write(',');
-                    WriteString(name, output);
+                    JsonWriter.WriteString(name, output);
                     output.Write(':');
-                    WriteLiteral(property.Value, output);
+                    JsonWriter.WriteLiteral(property.Value, output);
                 }
             }
         }
@@ -208,154 +188,6 @@ namespace NLog.Targets.Seq
                 san = '@' + san;
             }
             return san;
-        }
-        
-        static void WriteLiteral(object value, TextWriter output)
-        {
-            if (value == null)
-            {
-                output.Write("null");
-                return;
-            }
-
-            // Attempt to convert the object (if a string) to its literal type (int/decimal/date)
-            value = GetValueAsLiteral(value);
-
-            Action<object, TextWriter> writer;
-            if (LiteralWriters.TryGetValue(value.GetType(), out writer))
-            {
-                writer(value, output);
-                return;
-            }
-
-            WriteString(value.ToString(), output);
-        }
-
-        static void WriteToString(object number, TextWriter output)
-        {
-            output.Write(number.ToString());
-        }
-
-        static void WriteBoolean(bool value, TextWriter output)
-        {
-            output.Write(value ? "true" : "false");
-        }
-
-        static void WriteOffset(DateTimeOffset value, TextWriter output)
-        {
-            output.Write("\"");
-            output.Write(value.ToString("o"));
-            output.Write("\"");
-        }
-
-        static void WriteDateTime(DateTime value, TextWriter output)
-        {
-            output.Write("\"");
-            output.Write(value.ToString("o"));
-            output.Write("\"");
-        }
-
-        static void WriteString(string value, TextWriter output)
-        {
-            var content = Escape(value);
-            output.Write("\"");
-            output.Write(content);
-            output.Write("\"");
-        }
-
-        static string Escape(string s)
-        {
-            if (s == null) return null;
-
-            StringBuilder escapedResult = null;
-            var cleanSegmentStart = 0;
-            for (var i = 0; i < s.Length; ++i)
-            {
-                var c = s[i];
-                if (c < (char)32 || c == '\\' || c == '"')
-                {
-
-                    if (escapedResult == null)
-                        escapedResult = new StringBuilder();
-
-                    escapedResult.Append(s.Substring(cleanSegmentStart, i - cleanSegmentStart));
-                    cleanSegmentStart = i + 1;
-
-                    switch (c)
-                    {
-                        case '"':
-                            {
-                                escapedResult.Append("\\\"");
-                                break;
-                            }
-                        case '\\':
-                            {
-                                escapedResult.Append("\\\\");
-                                break;
-                            }
-                        case '\n':
-                            {
-                                escapedResult.Append("\\n");
-                                break;
-                            }
-                        case '\r':
-                            {
-                                escapedResult.Append("\\r");
-                                break;
-                            }
-                        case '\f':
-                            {
-                                escapedResult.Append("\\f");
-                                break;
-                            }
-                        case '\t':
-                            {
-                                escapedResult.Append("\\t");
-                                break;
-                            }
-                        default:
-                            {
-                                escapedResult.Append("\\u");
-                                escapedResult.Append(((int)c).ToString("X4"));
-                                break;
-                            }
-                    }
-                }
-            }
-
-            if (escapedResult != null)
-            {
-                if (cleanSegmentStart != s.Length)
-                    escapedResult.Append(s.Substring(cleanSegmentStart));
-
-                return escapedResult.ToString();
-            }
-
-            return s;
-        }
-
-        /// <summary>
-        /// GetValueAsLiteral attempts to transform the (string) object into a literal type prior to json serialization.
-        /// </summary>
-        /// <param name="value">The value to be transformed/parsed.</param>
-        /// <returns>A translated representation of the literal object type instead of a string.</returns>
-        static object GetValueAsLiteral(object value)
-        {
-            var str = value as string;
-            if (str == null)
-                return value;
-
-            // All number literals are serialized as a decimal so ignore other number types.
-            decimal decimalBuffer;
-            if (decimal.TryParse(str, out decimalBuffer))
-                return decimalBuffer;
-
-            // Standardize on dates if/when possible.
-            DateTime dateBuffer;
-            if (DateTime.TryParse(str, out dateBuffer))
-                return dateBuffer;
-
-            return value;
         }
     }
 }
