+++ src/NLog.Targets.Seq/project.json
@@ -32,7 +32,7 @@
   },
 
   "dependencies": {
-    "NLog": "4.5-alpha01"
+    "NLog": "4.5-alpha02"
   },
 
   "frameworks": {
